requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
APScheduler==3.10.4
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
selenium==4.15.2
